Check if debug.log exists

+++ src/debug.cpp
@@ -751,7 +751,8 @@ void DebugFile::init( DebugOutput output_mode, const cata_path &filename )
             this->filename = filename;
             // Continue with the old log file if it's smaller than 1 MiB
             namespace fs = std::filesystem;
-            if( fs::file_size( fs::path( filename ) ) >= 1024 * 1024 ) {
+            if( fs::exists( fs::path( filename ) )
+                && fs::file_size( fs::path( filename ) ) >= 1024 * 1024 ) {
                 std::error_code ec;
                 fs::rename( fs::path( filename ), fs::path( oldfile ), ec );
                 rename_failed = bool( ec );
